chore: rename Lval.mts such that its casing matches
