+++ rt/src/RawTuple.mts
@@ -1,5 +1,5 @@
 import { Level } from './Level.mjs';
-import { LVal, listStringRep } from './Lval.mjs';
+import { LVal, listStringRep } from './LVal.mjs';
 import { TroupeAggregateRawValue } from './TroupeRawValue.mjs';
 import { TroupeType } from './TroupeTypes.mjs';
 import * as levels from './Level.mjs'
