+++ rt/src/UnitVal.mts
@@ -1,6 +1,6 @@
 import RawUnit from './RawUnit.mjs'
 import { BOT } from './Level.mjs';
-import { LVal } from './Lval.mjs';
+import { LVal } from './LVal.mjs';
 
 export let __unit = new LVal (RawUnit, BOT, BOT)
