+++ rt/src/builtins/listToTuple.mts
@@ -1,6 +1,6 @@
 'use strict'
 import { UserRuntimeZero, Constructor, mkBase } from './UserRuntimeZero.mjs'
-import { LVal } from '../Lval.mjs';
+import { LVal } from '../LVal.mjs';
 import { assertIsList } from '../Asserts.mjs'
 import { mkTuple } from '../ValuesUtil.mjs'
 import { lub } from '../Level.mjs';
