+++ rt/src/builtins/string.mts
@@ -1,5 +1,5 @@
 import {UserRuntimeZero, Constructor, mkBase} from './UserRuntimeZero.mjs'
-import { LVal } from '../Lval.mjs';
+import { LVal } from '../LVal.mjs';
 import { assertIsString, assertIsNTuple, assertIsNumber } from '../Asserts.mjs'
 import { lub } from '../Level.mjs';
